% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chai_time.R
\name{chai_time}
\alias{chai_time}
\title{Find out whether you can have chai based on Indian snack preference and correct animal (subjective)}
\usage{
chai_time(food, animal)
}
\arguments{
\item{food}{A string vector with different popular Indian snacks}

\item{animal}{A string vector of animals}
}
\value{
A printed statement
}
\description{
Find out whether you can have chai based on Indian snack preference and correct animal (subjective)
}
